refactor(API): Rename 'db' namespace to 'database'
Fixes #221

BREAKING CHANGE:

Rename 'db' namespace to 'database'
