+++ nixnet/_session/base.py
@@ -354,7 +354,7 @@ class SessionBase(object):
             sched_index(int): Index to the schedule table that the LIN master executes.
 
             The schedule tables are sorted the way they are returned from the database
-            with the `nixnet.db._cluster.Cluster.lin_schedules` property.
+            with the `nixnet.database._cluster.Cluster.lin_schedules` property.
         """
         _funcs.nx_write_state(self._handle, constants.WriteState.LIN_SCHEDULE_CHANGE, _ctypedefs.u32(sched_index))
