+++ nixnet/database/_signal.py
